1. 增加支持redis注册中心 2. 注册中心代码优化

+++ kin-rpc-cluster/kin-rpc-cluster-api/src/main/java/org/kin/kinrpc/cluster/ClusterInvoker.java
@@ -31,10 +31,10 @@ abstract class ClusterInvoker<I> implements Closeable {
 
     private final Cluster cluster;
     private final int retryTimes;
-    private final int retryTimeout;
+    private final long retryTimeout;
     private final Url url;
 
-    public ClusterInvoker(Cluster cluster, int retryTimes, int retryTimeout, Url url) {
+    public ClusterInvoker(Cluster cluster, int retryTimes, long retryTimeout, Url url) {
         this.cluster = cluster;
         this.retryTimes = retryTimes;
         this.retryTimeout = retryTimeout;
